requirements.txt
# CHROMADB COMPATIBLE - Works with Python 3.11
# This version uses ChromaDB with all-MiniLM-L6-v2 with PyTorch support for re-ranker

# Core LangGraph and LangChain
langgraph>=0.0.55
langchain>=1.0.3
langchain-core>=1.0.0
langchain-community>=0.4.1
langchain-groq>=1.0.0
langchain-text-splitters>=0.0.1
langchain-chroma>=1.0.0
langchain-huggingface>=1.0.1

# LLM and API clients
groq>=0.4.2

# Vector Database with compatible embeddings
chromadb>=0.4.22

# Essential utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# Web Framework
flask>=3.0.0

# Data Processing
numpy>=1.24.0

# Web Scraping (for knowledge base updates)
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
schedule>=1.2.0

# Multi-Format File Support (Optional - for enhanced knowledge base)
# Install with: pip install -r requirements.txt
# Or selectively: pip install PyPDF2 python-docx pandas openpyxl markdown

# PDF Support
PyPDF2>=3.0.0

# Word Document Support
python-docx>=1.0.0

# CSV/Excel Support
pandas>=2.0.0
openpyxl>=3.0.0

# Markdown Support (enhanced parsing)
markdown>=3.4.0

# Re-ranker Support (Optional - for improved RAG relevance)
# NOW ENABLED: Using Python 3.11 with compatible dependencies
# Requires numpy<2 for PyTorch compatibility
sentence-transformers>=2.2.0
torch>=2.0.0
numpy<2

# Testing
pytest>=7.4.0